backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.1
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
numpy==1.26.2
numpy-financial==1.0.0
email-validator==2.1.0.post1
typing-extensions==4.9.0
asyncpg==0.29.0
pytest==8.2.0
pytest-asyncio==0.25.0
pyxirr
gotrue==2.8.0

pydantic-settings==2.1.0

# Email service libraries (choose one based on your preferred email provider)
# sendgrid==6.10.0  # Uncomment if using SendGrid for email service
# boto3==1.34.69  # Uncomment if using AWS SES for email service
# requests==2.31.0  # General-purpose HTTP library often used with email APIs